[660]: voxels more filters
