+++ src/modules/land-works/components/lands-explore-filters-modal/types.d.ts
@@ -1,9 +1,13 @@
 export interface MoreFiltersType {
   type: string;
-  size: minMax;
-  distanceToRoad: minMax;
-  distanceToPlaza: minMax;
-  distanceToDistrict: minMax;
+  size?: minMax;
+  distanceToRoad?: minMax;
+  distanceToPlaza?: minMax;
+  distanceToDistrict?: minMax;
+  height?: minMax;
+  area?: minMax;
+  hasBasement?: boolean;
+  isWaterFront?: boolean;
 }
 
 type minMax = {
